requirements.txt
python-shogi>=1.1.1
